refactor(cli): rename 'templates' group to 'template' and 'generate-docs' to 'docs'
