+++ cmd/template/publish.go
@@ -1,4 +1,4 @@
-package templates
+package template
 
 import (
 	"bytes"
